Apply review feedback to the polygon foundation
- Merge touching features instead of rejecting them: a FeatureCollection
  of adjacent regions (neighboring districts) previously tripped the
  self-intersection rejection through no fault of the input. Features
  are now validated individually, then dissolved together, with an
  edge-sharing test. (Review: daniellekain)
- Remove the polygons waffle switch and its migration, admin gating,
  and tests. (Review: protsack-stephan, daniellekain)
- Move PolygonForm/PolygonAdmin into api/models/polygon_admin.py,
  matching the partner_field_admin.py precedent, to slim admin.py.
  (Review: protsack-stephan)
- Cap GeoJSON file uploads at 10 MB with a friendly error.
  (Review: daniellekain)
- whitelist -> allowlist wording. (Review: daniellekain)
