+++ doc/release/RELEASE-NOTES.md
@@ -16,7 +16,6 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 * `0225_add_is_anonymized_to_source.py` - Adds the `Source.is_anonymized` boolean column (default `False`) and replaces the facility index functions `index_contributors`, `index_created_from_info`, `index_contributors_id`, `index_item_sectors`, `index_facility_names`, `index_facility_addresses`, `index_facility_list_items`, and `index_extended_fields` so anonymized sources keep their contributed data indexed while their contributor attribution is hidden. No backfill or reindex is required: no existing source is anonymized, the recomputed attribution values are unchanged for existing rows, and serializers read the new JSON keys with safe defaults. See OSDEV-3142.
 * `0226_add_slc_submission_quality_check_switch.py` - Adds the `slc_submission_quality_check` waffle switch (created active, so behavior is unchanged on deploy) that controls the LLM-backed SLC submission quality check. See OSDEV-3118.
 * `0227_create_polygon.py` - Adds the `Polygon` table: a reusable, named geographic boundary (an auto-generated unique `uuid` for joining, unique variable-style `name` — letters, digits, and underscores only — an optional human-friendly `display_name`, `description`, and a `geom` `MultiPolygonField` in EPSG:4326) that is not tied to country borders. Polygons may be any shape, including multiple disjoint parts and polygons with holes. `Polygon.get_production_locations()` returns the OS Hub production locations within a given boundary via the existing `location__within` PostGIS lookup, optionally narrowed by a whitelist of core-field filters (e.g. `{'country': ['IN'], 'sector': ['Apparel']}`) applied in the same query. Boundaries are added through the Django admin by pasting or uploading GeoJSON (a bare `Polygon`/`MultiPolygon`, a `Feature`, or a `FeatureCollection`, which are parsed, merged, and validated before saving); GeoJSON that declares a non-WGS-84 coordinate system, or whose coordinates fall outside valid longitude/latitude ranges, is rejected with a clear form error. This is foundational infrastructure for future boundary-driven features (e.g. per-region partner fields) rather than a user-facing change on its own; related context in [OSDEV-3125](https://opensupplyhub.atlassian.net/browse/OSDEV-3125).
-* `0228_add_polygons_switch.py` - Adds the `polygons` waffle switch (off by default), which gates the polygon section of the Django admin.
 
 #### Schema changes
 * [OSDEV-3142](https://opensupplyhub.atlassian.net/browse/OSDEV-3142) - Added the `is_anonymized` column to `api_source`.
